+++ src/plugins/braintree-dropin-error-mixin.js
@@ -3,9 +3,11 @@ import * as Sentry from '@sentry/vue';
 export default {
 	methods: {
 		processBraintreeDropInError(trackCategory, kivaBraintreeResponse) {
-			// eslint-disable-next-line max-len
-			const errorCode = kivaBraintreeResponse.errors?.[0]?.extensions?.code || kivaBraintreeResponse.errors?.[0]?.error;
-			const errorMessage = kivaBraintreeResponse.errors?.[0]?.message;
+			// extract error response
+			const errorObj = kivaBraintreeResponse?.[0] ?? kivaBraintreeResponse?.errors?.[0] ?? kivaBraintreeResponse;
+			// extract error code and message
+			const errorCode = errorObj?.extensions?.code ?? errorObj?.error ?? 'UNKNOWN_ERROR_CODE';
+			const errorMessage = errorObj?.message ?? JSON.stringify(kivaBraintreeResponse);
 			// eslint-disable-next-line max-len
 			const standardError = 'There was an error processing your payment. Please confirm your payment details and try again or contact our support team at <a href="mailto:contactus@kiva.org">contactus@kiva.org</a>.';
 
@@ -15,7 +17,7 @@ export default {
 			this.$kvTrackEvent(trackCategory, 'DropIn Payment Error', `${errorCode}: ${errorMessage}`);
 
 			// Log validation errors
-			Sentry.captureException(`${errorCode}:${errorMessage}`);
+			Sentry.captureException(`${errorCode}: ${errorMessage}`);
 		},
 	},
 };